+++ src/main/java/com/example/pigonair/domain/member/dto/MemberRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.pigonair.member.dto;
+package com.example.pigonair.domain.member.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
